asuka/requirements.txt
aiohttp==3.7.4.post0
python-dotenv==0.17.1
psycopg2==2.8.6